requirements.txt
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.58.0
imageio-ffmpeg==0.6.0
pypdf==6.7.0
Telethon==1.41.2
